backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib
python-multipart
python-dotenv
langchain
openai
psycopg2-binary
pydantic[email]
typing-extensions
pymongo==4.6.1
certifi==2023.7.22
bcrypt==4.0.1
PyJWT
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
dnspython==2.3.0
openai
python-multipart
pdf2image
pytesseract
pillow
